+++ src/server/devices/baseDeviceInterfaces/iBatteryDevice.ts
@@ -1,5 +1,9 @@
 import { iBaseDevice } from './iBaseDevice';
 
 export interface iBatteryDevice extends iBaseDevice {
+  /**
+   * The battery status of the device in percentage
+   * @type {number}
+   */
   battery: number;
 }