+++ src/server/devices/hmIPDevices/hmIpTaster.ts
@@ -38,7 +38,7 @@ export class HmIpTaster extends HmIPDevice implements iButtonSwitch, iBatteryDev
       case '0':
         switch (idSplit[4]) {
           case 'OPERATING_VOLTAGE':
-            this.battery = ((state.val as number) - 1.8) / 1.2;
+            this.battery = 100 * (((state.val as number) - 1.8) / 1.2);
             break;
         }
         break;